pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "konnect"
version = "0.1.0"
description = "Prusa Connect integration for generic Klipper/Moonraker printers"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "AGPL-3.0-or-later" }
authors = [{ name = "konnect contributors" }]
dependencies = [
    # Prusa's PyPI-published 0.8.1 predates the SET_VALUE command (added
    # in upstream master on 2024-09-23, commit ea6b14c) which is what
    # Connect uses for chamber temperature control on HT90-class
    # printers. Pinning to a specific commit on master gives us
    # SET_VALUE and reproducibility without tracking a moving target.
    "prusa.connect.sdk.printer @ git+https://github.com/prusa3d/Prusa-Connect-SDK-Printer.git@352caef1f1",
    "flask>=2.0",
    "websocket-client>=1.5",
    "requests>=2.28",
    "qrcode[pil]>=7.4",
]

[project.optional-dependencies]
nm = ["sdbus-networkmanager"]

[project.scripts]
konnect = "konnect.__main__:main"

[tool.setuptools.packages.find]
include = ["konnect*"]

[tool.setuptools.package-data]
konnect = ["static/*", "static/**/*"]
